requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
mysql-connector-python>=8.0.0
requests>=2.25.0
python-multipart>=0.0.5
pandas>=1.5.0
python-dotenv>=0.19.0
